refactor continued
