+++ bcipy/display/rsvp/mode/free_spell.py
@@ -1,5 +1,5 @@
 from psychopy import visual
-from bcipy.display.rsvp.rsvp_display import RSVPDisplay
+from bcipy.display.rsvp.display import RSVPDisplay
 from bcipy.helpers.stimuli import resize_image
 from bcipy.helpers.task import SPACE_CHAR
